requirements/base.txt
Django==1.8.4
djangorestframework==3.2.3
psycopg2==2.6.1
redis==2.10.3
